refactored some code based on code review
