+++ App.tsx
@@ -4,6 +4,7 @@ import { AppLoading } from 'expo';
 import * as Font from 'expo-font';
 import { StyleSheet, View } from 'react-native';
 import MealListScreen from './src/screens/MealList/MealListScreen';
+import RecipeScreen from './src/screens/Recipe/RecipeScreen';
 
 export default class App extends React.Component {
   state = {
@@ -31,8 +32,8 @@ export default class App extends React.Component {
     }
 
     return (
-      <View>
-        <MealListScreen />
+      <View style={styles.container}>
+        <RecipeScreen />
       </View>
     );
   }
