+++ package.json
@@ -9,7 +9,7 @@
   },
   "dependencies": {
     "expo": "~37.0.3",
-    "firebase": "7.9.0",
+    "firebase": "^7.15.1",
     "native-base": "^2.13.12",
     "react": "~16.9.0",
     "react-dom": "~16.9.0",
