+++ src/screens/Recipe/RecipeStep/styles.tsx
@@ -1,30 +1,24 @@
 import { StyleSheet } from 'react-native';
 
 const styles: any = StyleSheet.create({
-
     cardWrapper: {
+        alignSelf: 'center',
         width: '90%',
         borderRadius: 16,
-        alignSelf: 'center'
+        marginLeft: 0,
+        marginRight: 0,
     },
-    mealImage: {
-        height: 200,
-        width: '100%',
-        flex: 1,
-        borderTopLeftRadius: 16,
-        borderTopRightRadius: 16
-    },
-    cardFooter: {
+    steps: {
         flexDirection: 'column',
         borderRadius: 16
     },
-    mealName: {
+    stepHeader: {
         fontSize: 32,
         fontFamily: 'transat-text-bold',
         alignSelf: 'flex-start',
         paddingBottom: 16
     },
-    mealDescription: {
+    stepDescription: {
         fontSize: 24,
         fontFamily: 'transat-text-standard',
         alignSelf: 'flex-start',
