+++ src/screens/Recipe/styles.tsx
@@ -2,7 +2,10 @@ import { StyleSheet } from 'react-native';
 
 const styles: any = StyleSheet.create({
     container: {
-    }
+        flex: 1,
+        alignItems: 'center',
+        backgroundColor: '#F5F5F5'
+    },
 });
 
 export default styles;
